+++ scout/server/blueprints/cases/views.py
@@ -48,7 +48,7 @@ def cases(institute_id):
     all_cases = store.cases(institute_id, name_query=query, skip_assigned=skip_assigned)
     data = controllers.cases(store, all_cases, limit)
 
-    sanger_unevaluated = controllers.get_sanger_unevaluated(store, institute_id)
+    sanger_unevaluated = controllers.get_sanger_unevaluated(store, institute_id, current_user.email)
     if len(sanger_unevaluated)> 0:
         data['sanger_unevaluated'] = sanger_unevaluated